Adding Period::moveStartDate and Period::moveEndDate
To complement Period::move we are adding two other method

- Period::moveStartDate
- Period::moveEndDate

Both methods move the datepoint according to a given interval.
Depending on the DateInterval invert property, the interval is
either added or substracted from the datepoint.

We consequently deprecate

- Period::add
- Period::sub

which were at most misleading as they both represent a subset of Period::moveEndDate.